Making things a little cleaner. Internal data is more consistent (properly using and returning BOOLs), using int's instead of NSUIntegers, which were breaking NSUInteger->float conversion for display. Pass the entire flower object ptr to the detailView rather than breaking bits and pieces of it apart and passing those. Pull the getColorsByType from the FlowerContainer directly.

+++ FlowerPicker/FlowerPicker/Flower.h
@@ -15,8 +15,8 @@
 -(NSString*)type;
 -(NSString*)color;
 -(NSString*)imageName;
--(NSUInteger)dozCost;
--(NSUInteger)boqCost;
+-(int)dozCost;
+-(int)boqCost;
 -(NSDictionary*)season;
 -(BOOL)season: (NSString*)s; // Overload to allow getting specific season
 
@@ -25,8 +25,8 @@
 -(void)setType:(NSString*)t;
 -(void)setColor:(NSString*)c;
 -(void)setImageName:(NSString*)img;
--(void)setDozCost:(NSUInteger)dc;
--(void)setBoqCost:(NSUInteger)bc;
+-(void)setDozCost:(int)dc;
+-(void)setBoqCost:(int)bc;
 -(void)setSeason:(NSDictionary*)s;
 
 @end
